+++ internal/tests/api/targets/classification_test.go
@@ -92,7 +92,8 @@ func TestTargets(t *testing.T) {
 					},
 					Attrs: &pb.Target_TcpTargetAttributes{
 						TcpTargetAttributes: &pb.TcpTargetAttributes{
-							DefaultPort: &wrapperspb.UInt32Value{Value: 26},
+							DefaultPort:       &wrapperspb.UInt32Value{Value: 26},
+							DefaultClientPort: &wrapperspb.UInt32Value{Value: 27},
 						},
 					},
 					AuthorizedActions: []string{"action-1"},
@@ -153,7 +154,8 @@ func TestTargets(t *testing.T) {
 					},
 					Attrs: &pb.Target_TcpTargetAttributes{
 						TcpTargetAttributes: &pb.TcpTargetAttributes{
-							DefaultPort: &wrapperspb.UInt32Value{Value: 26},
+							DefaultPort:       &wrapperspb.UInt32Value{Value: 26},
+							DefaultClientPort: &wrapperspb.UInt32Value{Value: 27},
 						},
 					},
 					AuthorizedActions: []string{"action-1"},